OrderFromCartDraft to interface. #942
