+++ sphere-models/src/main/java/io/sphere/sdk/orders/OrderFromCartDraft.java
@@ -1,49 +1,28 @@
 package io.sphere.sdk.orders;
 
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.carts.Cart;
-import io.sphere.sdk.models.Base;
 import io.sphere.sdk.models.Versioned;
 
 import javax.annotation.Nullable;
 
-public class OrderFromCartDraft extends Base {
-    private final String id;
-    private final Long version;
-    @Nullable
-    private final String orderNumber;
-    @Nullable
-    private final PaymentState  paymentState;
+@JsonDeserialize(as = OrderFromCartDraftImpl.class)
+public interface OrderFromCartDraft {
+    String getId();
 
-    private OrderFromCartDraft(final String id, final Long version, @Nullable final String orderNumber, @Nullable final PaymentState paymentState) {
-        this.id = id;
-        this.version = version;
-        this.orderNumber = orderNumber;
-        this.paymentState = paymentState;
-    }
-
-    public static OrderFromCartDraft of(final Versioned<Cart> cart, @Nullable final String orderNumber, @Nullable final PaymentState paymentState) {
-        return new OrderFromCartDraft(cart.getId(), cart.getVersion(), orderNumber, paymentState);
-    }
-
-    public static OrderFromCartDraft of(final Versioned<Cart> cart) {
-        return new OrderFromCartDraft(cart.getId(), cart.getVersion(), null, null);
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public Long getVersion() {
-        return version;
-    }
+    Long getVersion();
 
     @Nullable
-    public String getOrderNumber() {
-        return orderNumber;
-    }
+    String getOrderNumber();
 
     @Nullable
-    public PaymentState getPaymentState() {
-        return paymentState;
+    PaymentState getPaymentState();
+
+    static OrderFromCartDraft of(final Versioned<Cart> cart, @Nullable final String orderNumber, @Nullable final PaymentState paymentState) {
+        return new OrderFromCartDraftImpl(cart.getId(), cart.getVersion(), orderNumber, paymentState);
+    }
+
+    static OrderFromCartDraft of(final Versioned<Cart> cart) {
+        return new OrderFromCartDraftImpl(cart.getId(), cart.getVersion(), null, null);
     }
 }
